Crate gfx_traits

Modules

print_tree

Structs

ByteIndex

An index that refers to a byte offset in a text run. This could point to the middle of a glyph.

Epoch

A newtype struct for denoting the age of messages; prevents race conditions.

StackingContextId

A unique ID for every stacking context.

Enums

FontData

FragmentType

The type of fragment that a scroll root is created for.

Constants

SPECIAL_SCROLL_ROOT_ID_MASK

If none of the bits outside this mask are set, the scroll root is a special scroll root. Note that we assume that the top 16 bits of the address space are unused on the platform.

Statics

NEXT_SPECIAL_SCROLL_ROOT_ID

The next ID that will be used for a special scroll root id.

Traits

WebrenderApi

Functions

combine_id_with_fragment_type

next_special_id

Returns a new scroll root ID for a scroll root.

node_id_from_scroll_id