Trait gfx_traits::paint_listener::PaintListener

pub trait PaintListener {
    fn native_display(&mut self) -> Option<NativeDisplay>;
    fn initialize_layers_for_pipeline(&mut self, pipeline_id: PipelineId, properties: Vec<LayerProperties>, epoch: Epoch);
    fn assign_painted_buffers(&mut self, pipeline_id: PipelineId, epoch: Epoch, replies: Vec<(LayerId, Box<LayerBufferSet>)>, frame_tree_id: FrameTreeId);
    fn ignore_buffer_requests(&mut self, buffer_requests: Vec<BufferRequest>);
    fn notify_paint_thread_exiting(&mut self, pipeline_id: PipelineId);
}

The interface used by the painter to acquire draw targets for each paint frame and submit them to be drawn to the display.

Required Methods

fn native_display(&mut self) -> Option<NativeDisplay>

fn initialize_layers_for_pipeline(&mut self, pipeline_id: PipelineId, properties: Vec<LayerProperties>, epoch: Epoch)

Informs the compositor of the layers for the given pipeline. The compositor responds by creating and/or destroying paint layers as necessary.

fn assign_painted_buffers(&mut self, pipeline_id: PipelineId, epoch: Epoch, replies: Vec<(LayerId, Box<LayerBufferSet>)>, frame_tree_id: FrameTreeId)

Sends new buffers for the given layers to the compositor.

fn ignore_buffer_requests(&mut self, buffer_requests: Vec<BufferRequest>)

Inform the compositor that these buffer requests will be ignored.

fn notify_paint_thread_exiting(&mut self, pipeline_id: PipelineId)