Struct gfx_traits::LayerProperties

pub struct LayerProperties {
    pub id: LayerId,
    pub parent_id: Option<LayerId>,
    pub rect: Rect<f32>,
    pub background_color: Color,
    pub scroll_policy: ScrollPolicy,
    pub transform: Matrix4D<f32>,
    pub perspective: Matrix4D<f32>,
    pub subpage_pipeline_id: Option<PipelineId>,
    pub establishes_3d_context: bool,
    pub scrolls_overflow_area: bool,
}

All layer-specific information that the painting task sends to the compositor other than the buffer contents of the layer itself.

Fields
id: LayerId

An opaque ID. This is usually the address of the flow and index of the box within it.

parent_id: Option<LayerId>

The id of the parent layer.

rect: Rect<f32>

The position and size of the layer in pixels.

background_color: Color

The background color of the layer.

scroll_policy: ScrollPolicy

The scrolling policy of this layer.

transform: Matrix4D<f32>

The transform for this layer

perspective: Matrix4D<f32>

The perspective transform for this layer

subpage_pipeline_id: Option<PipelineId>

The subpage that this layer represents. If this is Some, this layer represents an iframe.

establishes_3d_context: bool

Whether this layer establishes a new 3d rendering context.

scrolls_overflow_area: bool

Whether this layer scrolls its overflow area.